+++ admin/tabs/monitor/MonitorResultsDisplay.js
@@ -7,13 +7,16 @@
 import {hbox} from '@xh/hoist/cmp/layout';
 import {hoistCmp} from '@xh/hoist/core';
 import {tile} from './Tile';
+import {isEmpty} from 'lodash';
 
 export const monitorResultsDisplay = hoistCmp.factory(
-    ({model}) => hbox({
-        className: 'xh-monitor-status-display',
-        items: model.results.map((check, idx) => tile({
-            key: `tile-${idx}`,
-            check
-        }))
-    })
+    ({model}) => {
+        const {results} = model;
+        return hbox({
+            className: 'xh-monitor-status-display',
+            items: isEmpty(results) ?
+                'No monitors configured for this application.' :
+                model.results.map((check, idx) => tile({key: `tile-${idx}`, check}))
+        });
+    }
 );