+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## v36.0.0-SNAPSHOT - under development
+## v36.0.0-rc2 - 2020-09-02
+
+### 🎁 New Features
+
+* Continued Usability improvements to FilterChooser
+* The Admin console's Monitor Status tab displays more clearly when there are no active monitors.
+
+[Commit Log](https://github.com/xh/hoist-react/compare/v36.0.0-rc1...v36.0.0-rc2)
+
+## v36.0.0-rc1 - 2020-08-31
 
 ### 🎁 New Features
 
@@ -125,7 +134,7 @@ below regarding related updates to `GridModel.columns` config processing.
 * mobx-react `6.2 -> 6.3`
 * store2 `2.11 -> 2.12`
 
-[Commit Log](https://github.com/xh/hoist-react/compare/v35.2.1...develop)
+[Commit Log](https://github.com/xh/hoist-react/compare/v35.2.1...v36.0.0-rc1)
 
 
 ## v35.2.1 - 2020-07-31